// ExistingCode/src/data-access/domain/tables/assay/schema.ts
export const AssayTableInsertBaseSchema = z.object({
	/**
	 * dataType: uniqueidentifier
	 * defaultValue: (newid())
	 */
	assayId: z.any().optional(),
	/**
	 * dataType: nvarchar
	 */
	organization: z.string().max(30),
	/**
	 * dataType: uniqueidentifier
	 */
	sampleId: z.any(),
	/**
	 * dataType: nvarchar
	 */
	genericMethod: z.string().max(50),
	/**
	 * dataType: nvarchar
	 */
	labElement: z.string().max(50),
	/**
	 * dataType: nvarchar
	 */
	element: z.string().max(50),
	/**
	 * dataType: nvarchar
	 */
	assayClassification: z.string().max(50),
	/**
	 * dataType: nvarchar
	 */
	repeat: z.string().max(10),
	/**
	 * dataType: smallint
	 */
	preferred: z.number().int(),
	/**
	 * dataType: nvarchar
	 */
	originalMethod: z.string().max(50),
	/**
	 * dataType: nvarchar
	 */
	unitCode: z.string().max(50),
	/**
	 * dataType: nvarchar
	 */
	assayResult: z.string().max(20),
	/**
	 * dataType: decimal
	 */
	assayResultNum: z.number().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	labCode: z.string().max(20),
	/**
	 * dataType: nvarchar
	 */
	batchNo: z.string().max(50),
	/**
	 * dataType: decimal
	 */
	limitLower: z.number().nullable().optional(),
	/**
	 * dataType: decimal
	 */
	limitUpper: z.number().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	sysAssayStatus: z.string().max(4).nullable().optional(),
	/**
	 * dataType: decimal
	 */
	sysResult: z.number().nullable().optional(),
	/**
	 * dataType: smallint
	 */
	labSequence: z.number().int().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	passFail: z.string().max(1000).nullable().optional(),
	/**
	 * dataType: int
	 * defaultValue: ((0))
	 */
	sourceRowNumber: z.number().int().optional(),
	/**
	 * dataType: varchar
	 */
	reading: z.string().max(50).nullable().optional(),
	/**
	 * dataType: varbinary
	 */
	migrationRv: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	jsonData: z.string().max(-1).nullable().optional(),
	/**
	 * dataType: int
	 */
	migrationId: z.number().int().nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((0))
	 */
	reportIncludeInd: z.any().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	validationStatus: z.number().int().optional(),
	/**
	 * dataType: nvarchar
	 */
	validationErrors: z.string().max(4000).nullable().optional(),
	/**
	 * dataType: tinyint
	 * defaultValue: ((0))
	 */
	rowStatus: z.number().int().optional(),
	/**
	 * dataType: uniqueidentifier
	 */
	supersededById: z.any().nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((1))
	 */
	activeInd: z.any().optional(),
	/**
	 * dataType: datetimeoffset
	 * defaultValue: (sysdatetimeoffset())
	 */
	createdOnDt: z.any().optional(),
	/**
	 * dataType: nvarchar
	 * defaultValue: (CONVERT([nvarchar](50),coalesce(session_context(N'UserName'),suser_sname())))
	 */
	createdBy: z.string().max(50).optional(),
	/**
	 * dataType: datetimeoffset
	 */
	modifiedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	modifiedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: timestamp
	 */
	rv: z.date(),
});